refactor: use consistent case in file names
